Add files via upload
The individual login files to where the user can login as a homecook or as a patron. The styling for these pages follow login-1.css. 
The login.js file attaches an event listener to each element with the class .login-btn. When any of these buttons are clicked, it triggers an alert that displays a message saying "Login as" followed by the button's text content, trimmed of any leading or trailing whitespace (updates to be made).