Add files via upload
These are the files for the front-end functionality of the search bar function. 
It's currently linking to a recipe API from the web (will update to update current db). 

This JavaScript code interacts with a web API to fetch recipe data, which it then displays on a webpage. Users can dynamically filter the displayed recipes in real-time by typing into a search bar, because of an event listener on the search input that updates the display as the user types. The code handles fetching, filtering, and rendering of the API data entirely on the client side.